+++ tests/ui/blanket_bad_depth.rs
@@ -1,0 +1,6 @@
+use batch_impl::batch_impl;
+
+#[batch_impl(#blanket(#all){Box:abc})]
+trait BadDepth {
+    fn m(&self) -> u32;
+}