+++ tests/ui/blanket_ptr.rs
@@ -1,0 +1,6 @@
+use batch_impl::batch_impl;
+
+#[batch_impl(#blanket(#all){*const})]
+trait BadPtr {
+    fn m(&self) -> u32;
+}