+++ tests/ui/const_cycle.rs
@@ -1,0 +1,7 @@
+use batch_impl::batch_trait;
+
+trait T {}
+batch_trait!(
+    @a=@a;
+    T: @a;
+);