+++ tests/ui/const_forward.rs
@@ -1,0 +1,8 @@
+use batch_impl::batch_trait;
+
+trait T {}
+batch_trait!(
+    @a=@b;
+    @b=[u8];
+    T: @a;
+);