+++ tests/ui/const_range_bad.rs
@@ -1,0 +1,4 @@
+use batch_impl::batch_impl;
+
+#[batch_impl(@u32..u8)]
+trait Bad {}